backend/pyproject.toml
[project]
name = "petpal-backend"
version = "0.1.0"
description = "PetPal - AI Pet Health Assistant Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "click>=8.1.0",
    "sse-starlette>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pyjwt>=2.9.0",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
    "litellm>=1.50.0,!=1.82.7,!=1.82.8",
    "python-multipart>=0.0.9",
    "pgvector>=0.3.0",
    "google-cloud-storage>=2.18.0",
    "app-store-server-library>=3.0.0",
    "jinja2>=3.1.0",
]

[project.scripts]
debug = "app.admin_cli.main:debug_shim"
admin = "app.admin_cli.main:cli"

[project.optional-dependencies]
dev = ["respx>=0.21.0", "pytest-asyncio>=0.23.0"]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
